[FIX] stock@saas-5: handle manual fields when merging stock.picking.{in,out}

+++ migrations/stock/7.saas~5.1.1/pre-10-nopicking-out-in.py
@@ -2,7 +2,6 @@ from openerp.addons.base.maintenance.migrations import util
 
 def migrate(cr, version):
 
-    #Views delete
     util.remove_record(cr, 'stock.view_picking_in_tree')
     util.remove_record(cr, 'stock.view_picking_in_form')
     util.remove_record(cr, 'stock.view_picking_in_calendar')
@@ -11,19 +10,44 @@ def migrate(cr, version):
     util.remove_record(cr, 'stock.view_picking_out_form')
     util.remove_record(cr, 'stock.view_picking_out_calendar')
 
-    #Change stock.picking.in/out to stock.picking
+    # Change stock.picking.in/out to stock.picking
     cr.execute("SELECT model, id FROM ir_model WHERE model IN %s",
-               (('stock.picking', 'stock.picking.in', 'stock.picking.out'),)
-    )
-    
+               [('stock.picking', 'stock.picking.in', 'stock.picking.out')])
+
     ids = dict(cr.fetchall())
     id_sp = ids['stock.picking']
     id_in = ids['stock.picking.in']
     id_out = ids['stock.picking.out']
-    
+
+    # change model of manual fields (if not already present)
+    for in_out in [id_in, id_out]:
+        cr.execute("""UPDATE ir_model_fields in_out
+                         SET model_id=%s
+                       WHERE model_id=%s
+                         AND manual=true
+                         AND NOT EXISTS(
+                            SELECT 1
+                              FROM ir_model_fields sp
+                             WHERE model_id=%s
+                               AND name=in_out.name
+                         )
+                   """, [id_sp, in_out, id_sp])
+
+    # remaining manual fields must be reassigned
+    cr.execute("""SELECT io.id, sp.id
+                    FROM ir_model_fields io, ir_model_fields sp
+                   WHERE io.model_id IN %s
+                     AND sp.model_id = %s
+                     AND io.manual = true
+                     AND io.name = sp.name
+               """, [(id_in, id_out), id_sp])
+    for o, n in cr.fetchall():
+        util.replace_record_references(cr, ('ir.model.fields', o), ('ir.model.fields', n))
+
     cr.execute("DELETE FROM ir_model_fields WHERE model_id IN %s RETURNING id", ((id_in, id_out),))
     fields_ids = tuple(x[0] for x in cr.fetchall())
-    cr.execute("DELETE FROM ir_model_data WHERE model=%s AND res_id IN %s", ('ir.model.fields', fields_ids))
+    if fields_ids:
+        cr.execute("DELETE FROM ir_model_data WHERE model=%s AND res_id IN %s", ('ir.model.fields', fields_ids))
     cr.execute("DELETE FROM ir_model_data WHERE model=%s AND res_id IN %s", ('ir.model', (id_in, id_out)))
     util.replace_record_references(cr, ('ir.model', id_in), ('ir.model', id_sp))
     util.replace_record_references(cr, ('ir.model', id_out), ('ir.model', id_sp))
